+++ src/data.c
@@ -6,7 +6,7 @@
 /*   By: svereten <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/16 10:03:52 by svereten          #+#    #+#             */
-/*   Updated: 2024/09/20 15:42:21 by svereten         ###   ########.fr       */
+/*   Updated: 2024/09/22 13:50:58 by svereten         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "fdf.h"
@@ -23,12 +23,11 @@ static t_data	*data_init(void)
 
 static void	data_free(t_data *data)
 {
-	size_t	i;
+	int	i;
 
 	if (!data)
 		return ;
 	i = 0;
-
 	while (data->matrix && i < data->actual_height)
 	{
 		if (data->matrix)
